doc: fix release guide example consistency
This changeset fixes two consistency issues in this single example that
guides releasers on how to push the tag to the remote repo.

The first issue is that the remote git reference is consistently
referred as `upstream` elsewhere in the guide.

The second issue being that the version example used in the rest of the
guide should be a fictional `v1.2.3` as stated in the "How to create a
release" guide notes:

>  * Examples will use the fictional release version `1.2.3`.

PR-URL: https://github.com/nodejs/node/pull/44385
Reviewed-By: Antoine du Hamel <duhamelantoine1995@gmail.com>
Reviewed-By: Richard Lau <rlau@redhat.com>
Reviewed-By: Rafael Gonzaga <rafael.nunu@hotmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>